+++ packages/terra-application-navigation/CHANGELOG.md
@@ -3,8 +3,8 @@ Changelog
 
 Unreleased
 ----------
-### Changed
-* Updated `content-layout` background color.
+### Fixed
+* Fixed `settings` api called when `help` clicked in drawer menu.
 
 1.18.0 - (February 4, 2020)
 ------------------